requirements.txt
Django == 5.0.3

celery~=5.3.6
python-dotenv~=1.0.1
django-filter~=23.5
APScheduler~=3.10.4
django-allauth~=0.61.0